refactor for golang suggested standards
