--- cmd/hammer_events/main.go
@@ -1,14 +1,0 @@
-package main
-
-import (
-	perun_services "main.go/services"
-	"main.go/utils"
-)
-
-func main() {
-
-	utils.Logger = utils.GetLogger(true, "", "eventslog")
-	synchronizer := perun_services.DockerSynchronizationService{}
-
-	synchronizer.Listen()
-}
