+++ cmd/logs.go
@@ -5,7 +5,7 @@ Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+	http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
@@ -16,8 +16,8 @@ limitations under the License.
 package cmd
 
 import (
+	"github.com/perun-cloud-inc/perunctl/services"
 	"github.com/spf13/cobra"
-	"main.go/services"
 )
 
 // logsCmd represents the logs command
@@ -25,24 +25,18 @@ var logsCmd = &cobra.Command{
 	Use:   "logs",
 	Short: "Get logs from a container",
 	Run: func(cmd *cobra.Command, args []string) {
-		tail, _ := cmd.Flags().GetBool("tail")
-		container_id, _ := cmd.Flags().GetString("id")
-		services.GetLogs(container_id, tail)
-
+		containerId, _ := cmd.Flags().GetString("id")
+		err := services.GetLogs(containerId)
+		if err != nil {
+			cobra.CheckErr(err)
+		}
 	},
 }
 
 func init() {
-	rootCmd.AddCommand(logsCmd)
-	logsCmd.Flags().BoolP("tail", "t", false, "Tail the logs")
+	RootCmd.AddCommand(logsCmd)
 	logsCmd.Flags().StringP("id", "i", "", "Container id")
-	// Here you will define your flags and configuration settings.
 
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// logsCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// logsCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	// TODO: Implement tail flag, and re-enable it
+	//logsCmd.Flags().BoolP("tail", "t", false, "Tail the logs")
 }
