+++ utils/constants.go
@@ -1,5 +1,6 @@
 package utils
 
-var PERUN_HOME = "/.perun/"
-var BIN_HOME = PERUN_HOME + "bin/"
-var WORKSPACES_HOME = PERUN_HOME + "workspaces/"
+const (
+	PerunHome      = "/.perun/"
+	WorkspacesHome = PerunHome + "workspaces/"
+)
